Distinguish between Diff and DiffForSingleRepository
