+++ src/test/java/de/uhd/ifi/se/decision/management/jira/git/gitclient/TestGetRationaleFromCommitMessages.java
@@ -8,7 +8,7 @@ import org.eclipse.jgit.lib.Ref;
 import org.eclipse.jgit.revwalk.RevCommit;
 import org.junit.Test;
 
-import de.uhd.ifi.se.decision.management.jira.git.model.Diff;
+import de.uhd.ifi.se.decision.management.jira.git.model.DiffForSingleRepository;
 import net.java.ao.test.jdbc.NonTransactional;
 
 public class TestGetRationaleFromCommitMessages extends TestSetUpGit {
@@ -16,7 +16,7 @@ public class TestGetRationaleFromCommitMessages extends TestSetUpGit {
 	@Test
 	@NonTransactional
 	public void nullOrEmptyFeatureBranchCommits() {
-		assertEquals(0, new Diff().getRationaleElementsFromCommitMessages().size());
+		assertEquals(0, new DiffForSingleRepository().getRationaleElementsFromCommitMessages().size());
 	}
 
 	@Test
@@ -24,7 +24,7 @@ public class TestGetRationaleFromCommitMessages extends TestSetUpGit {
 	public void fromFeatureBranchCommits() {
 		Ref featureBranch = gitClient.getRefs("TEST-4.feature.branch").get(0);
 		List<RevCommit> commits = gitClient.getCommits(featureBranch);
-		Diff diff = new Diff();
+		DiffForSingleRepository diff = new DiffForSingleRepository();
 		diff.setCommits(commits);
 		assertEquals(6, diff.getRationaleElementsFromCommitMessages().size());
 	}
